\name{svIDE-package}
\alias{svIDE-package}
\alias{svIDE}
\docType{package}

\title{ SciViews GUI API - IDE and code editor functions }
\description{
  Function for the GUI API to interact with external IDE/code editors.
}
\details{
  \tabular{ll}{
    Package: \tab svIDE\cr
    Type: \tab Package\cr
    Version: \tab 0.9-52\cr
    Date: \tab 2014-03-02\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Philippe Grosjean

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API IDE and code editor, language syntax }
